hello , most probably final!

+++ lmsApplication/src/main/java/com/example/lmsapplication/dto/AuditRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuditRepository extends JpaRepository<AuditDetails, Long> {
+public interface AuditRepository extends JpaRepository<AuditDetails, Integer> {
     AuditDetails findByLeave(Integer leave_Id);
 }
